moved cache into its own module, and the start of caching db calls
