+++ benwaonlineapi/util.py
@@ -2,36 +2,15 @@
 Contains any utility functions used by processors or the benwaonline frontend.
 """
 import os
-import json
 import requests
 from jose import jwt, exceptions
 from flask import current_app
 from flask_restless import ProcessingException
-from pymemcache.client.base import Client
-
 from benwaonlineapi.config import app_config
+from benwaonlineapi.cache import cache
 
 cfg = app_config[os.getenv('FLASK_CONFIG')]
 ALGORITHMS = ['RS256']
-
-def json_serializer(key, value):
-    if type(value) == str:
-        return value, 1
-    return json.dumps(value), 2
-
-def json_deserializer(key, value, flags):
-    if flags == 1:
-        return value.decode('utf-8')
-    if flags == 2:
-        return json.loads(value.decode('utf-8'))
-    raise Exception("Unknown serialization format")
-
-cache = Client(
-    (cfg.MEMCACHED_HOST, cfg.MEMCACHED_PORT),
-    connect_timeout=5,
-    serializer=json_serializer,
-    deserializer=json_deserializer
-)
 
 def verify_token(token, jwks, audience=cfg.API_AUDIENCE, issuer=cfg.ISSUER):
     unverified_header = jwt.get_unverified_header(token)
